Firebase Authentication with Google's One Tap Sign-In
This took a while because it is my first time using Kotlin, and a bit of the example code from Google's Developer Documentation was deprecated so I had to find alternatives.

I have made a temporary light pink theme/style for the app, but my final intentions are to have it dark themed.